+++ src/SqlVirtualMachine/SqlVirtualMachine/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed a bug of parameter `VirtualMachineResourceId` of cmdlet `New-AzSqlVM`.
 
 ## Version 2.1.0
 * Added more parameters on cmdlet `Update-AzSqlVM`.